==> notebooks/04 Model Training.ipynb <==
# 04 Entrenamiento de Modelos

En esta sección, vamos a entrenar varios modelos de machine learning para predecir si un cliente suscribirá un depósito a plazo. Probaremos diferentes algoritmos, ajustaremos sus hiperparámetros y evaluaremos su rendimiento.

## 1. Cargar Datos Preprocesados y Construidos

## 1. Cargar el Conjunto de Datos
# Importar la librería necesarias
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
import joblib

# Seleccionar el tipo de balanceo a utilizar
tipo_balanceo = 'oversampled' # 'original', 'oversampled', 'undersampled'

if tipo_balanceo == 'original':
    dir_path = '../data/interim'
elif tipo_balanceo == 'oversampled':
    dir_path = '../data/processed/oversampled'
elif tipo_balanceo == 'undersampled':
    dir_path = '../data/processed/undersampled'

# Cargar el conjunto de datos
X_train = pd.read_csv(f'{dir_path}/X_train.csv')
X_test = pd.read_csv(f'{dir_path}/X_test.csv')
y_train = pd.read_csv(f'{dir_path}/y_train.csv')
y_test = pd.read_csv(f'{dir_path}/y_test.csv')

# Mostrar los primeros registros de los datos de entrenamiento
X_train.head()

## 3. Entrenamiento de Modelos

### 3.1. Entrenamiento de un Modelo de Regresión Logística

from sklearn.linear_model import LogisticRegression

# Crear una instancia del modelo
lr = LogisticRegression(random_state=8)

# Entrenar el modelo
lr.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(lr, '../models/1_lr_model.pkl')

### 3.2. Entrenamiento de un Modelo de Árbol de Decisión

from sklearn.tree import DecisionTreeClassifier

# Crear una instancia del modelo
dt = DecisionTreeClassifier(random_state=8)

# Entrenar el modelo
dt.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(dt, '../models/2_dt_model.pkl')

### 3.3. Entrenamiento de un Modelo de Bosque Aleatorio

from sklearn.ensemble import RandomForestClassifier

# Crear una instancia del modelo
rf = RandomForestClassifier(random_state=8)

# Entrenar el modelo
rf.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(rf, '../models/3_rf_model.pkl')

### 3.4. Entrenamiento de un Modelo de Gradiente Boosting

from sklearn.ensemble import GradientBoostingClassifier

# Crear una instancia del modelo
gb = GradientBoostingClassifier(random_state=8)

# Entrenar el modelo
gb.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(gb, '../models/4_gb_model.pkl')

### 3.5. Entrenamiento de un Modelo de K-Nearest Neighbors

from sklearn.neighbors import KNeighborsClassifier

# Crear una instancia del modelo
knn = KNeighborsClassifier()

# Entrenar el modelo
knn.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(knn, '../models/5_knn_model.pkl')

### 3.6. Entrenamiento de un Modelo de Naive Bayes

from sklearn.naive_bayes import GaussianNB

# Crear una instancia del modelo
nb = GaussianNB()

# Entrenar el modelo
nb.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(nb, '../models/6_nb_model.pkl')

### 3.7. Entrenamiento de un Modelo de Red Neuronal

from sklearn.neural_network import MLPClassifier

# Crear una instancia del modelo
mlp = MLPClassifier(random_state=8)

# Entrenar el modelo
mlp.fit(X_train, y_train.values.ravel())

# Guardar el modelo
joblib.dump(mlp, '../models/7_mlp_model.pkl')

### 3.8. Entrenamiento de un Modelo de Support Vector Machine (demora mucho)

# from sklearn.svm import SVC

# Crear una instancia del modelo
# svc = SVC(probability=True, random_state=8)

# Entrenar el modelo
# svc.fit(X_train, y_train.values.ravel())

# Guardar el modelo
# joblib.dump(svc, '../models/8_svc_model.pkl')
