==> notebooks/05 Model Evaluation.ipynb <==
# 05 Evaluación de Modelos

## 1. Cargar Datos Procesados

## 1. Cargar el Conjunto de Datos
# Importar la librería necesarias
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

# Seleccionar el tipo de balanceo a utilizar
tipo_balanceo = 'oversampled' # 'original', 'oversampled', 'undersampled'

if tipo_balanceo == 'original':
    dir_path = '../data/interim'
elif tipo_balanceo == 'oversampled':
    dir_path = '../data/processed/oversampled'
elif tipo_balanceo == 'undersampled':
    dir_path = '../data/processed/undersampled'

# Cargar el conjunto de datos
X_test = pd.read_csv(f'{dir_path}/X_test.csv')
y_test = pd.read_csv(f'{dir_path}/y_test.csv')

# Mostrar los primeros registros de los datos de entrenamiento
X_test.head()

## 2. Cargar Modelos Preentrenados

import joblib

# Carga Modelos preentrenados
lr = joblib.load('../models/1_lr_model.pkl')
dt = joblib.load('../models/2_dt_model.pkl')
rf = joblib.load('../models/3_rf_model.pkl')
gb = joblib.load('../models/4_gb_model.pkl')
knn = joblib.load('../models/5_knn_model.pkl')
nb = joblib.load('../models/6_nb_model.pkl')
mlp = joblib.load('../models/7_mlp_model.pkl')
# svc = joblib.load('../models/8_svc_model.pkl')

## 3. Evaluacion de Modelos

### 3.1. Entrenamiento de un Modelo de Regresión Logística

# Hacer predicciones en el conjunto de prueba
y_pred_lr = lr.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_lr))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_lr))
print("Classification Report:\n", classification_report(y_test, y_pred_lr))

### 3.2. Entrenamiento de un Modelo de Árbol de Decisión

# Hacer predicciones en el conjunto de prueba
y_pred_dt = dt.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_dt))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_dt))
print("Classification Report:\n", classification_report(y_test, y_pred_dt))

### 3.3. Entrenamiento de un Modelo de Bosque Aleatorio

# Hacer predicciones en el conjunto de prueba
y_pred_rf = rf.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_rf))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_rf))
print("Classification Report:\n", classification_report(y_test, y_pred_rf))

### 3.4. Entrenamiento de un Modelo de Gradiente Boosting

# Hacer predicciones en el conjunto de prueba
y_pred_gb = gb.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_gb))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_gb))
print("Classification Report:\n", classification_report(y_test, y_pred_gb))

### 3.5. Entrenamiento de un Modelo de K-Nearest Neighbors

# Hacer predicciones en el conjunto de prueba
y_pred_knn = knn.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_knn))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_knn))
print("Classification Report:\n", classification_report(y_test, y_pred_knn))

### 3.6. Entrenamiento de un Modelo de Naive Bayes

# Hacer predicciones en el conjunto de prueba
y_pred_nb = nb.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_nb))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_nb))
print("Classification Report:\n", classification_report(y_test, y_pred_nb))

### 3.7. Entrenamiento de un Modelo de Red Neuronal

# Hacer predicciones en el conjunto de prueba
y_pred_mlp = mlp.predict(X_test)

# Evaluar el modelo
print("Accuracy:", accuracy_score(y_test, y_pred_mlp))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_mlp))
print("Classification Report:\n", classification_report(y_test, y_pred_mlp))

### 3.8. Entrenamiento de un Modelo de Support Vector Machine (demora mucho)

# Hacer predicciones en el conjunto de prueba
# y_pred_svc = svc.predict(X_test)

# Evaluar el modelo
# print("Accuracy:", accuracy_score(y_test, y_pred_svc))
# print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred_svc))
# print("Classification Report:\n", classification_report(y_test, y_pred_svc))
